This commit will add testing for Order model

+++ migrations/sqls/20230102234421-orders-table-up.sql
@@ -3,6 +3,6 @@ CREATE TABLE orders
     id SERIAL PRIMARY KEY,
     product_Id INTEGER REFERENCES products(id),
     quantity INTEGER NOT NULL,
-    user_id bigint REFERENCES users(id),
+    user_id INTEGER REFERENCES users(id),
     status VARCHAR(255) DEFAULT 'active'
 );
